Extract Rust's parsing of `--async` to a shared library
This should hopefully enable it to share the same parsing of this option
amongst all bindings generators.
