+++ crates/core/src/lib.rs
@@ -12,6 +12,8 @@ mod types;
 pub use types::{TypeInfo, Types};
 mod path;
 pub use path::name_package_module;
+mod async_;
+pub use async_::AsyncFilterSet;
 
 #[derive(Default, Copy, Clone, PartialEq, Eq, Debug)]
 pub enum Direction {
